# Conflicts:
#	Kandanda/Kandanda.BusinessLayer.Testing/GroupPhaseGeneratorTest.cs
#	Kandanda/Kandanda.BusinessLayer.Testing/ParticipantServiceTest.cs
#	Kandanda/Kandanda.BusinessLayer.Testing/TournamentServiceTest.cs
#	Kandanda/Kandanda.BusinessLayer/ServiceImplementations/PhaseService.cs
#	Kandanda/Kandanda.BusinessLayer/ServiceInterfaces/IPhaseService.cs
#	Kandanda/Kandanda.Dal/KandandaDbContext.cs

+++ Kandanda/Kandanda.BusinessLayer/ServiceImplementations/TournamentService.cs
@@ -1,14 +1,23 @@
 ﻿using System.Collections.Generic;
+using System.Data.Entity;
 using System.Linq;
+using System.Threading.Tasks;
 using Kandanda.BusinessLayer.PhaseGenerators;
 using Kandanda.BusinessLayer.ServiceInterfaces;
 using Kandanda.Dal;
-using Kandanda.Dal.DataTransferObjects;
+using Kandanda.Dal.Entities;
 
 namespace Kandanda.BusinessLayer.ServiceImplementations
 {
     public sealed class TournamentService : ServiceBase, ITournamentService
     {
+        private readonly IPhaseService _phaseService;
+
+        public TournamentService(KandandaDbContext dbContext) : base(dbContext)
+        {
+            _phaseService = new PhaseService(dbContext);
+        }
+
         public Tournament CreateEmpty(string name)
         {
             return Create(new Tournament
@@ -17,26 +26,30 @@ namespace Kandanda.BusinessLayer.ServiceImplementations
             });
         }
 
+        public async Task<List<Phase>> GetPhasesByTournamentAsync(Tournament tournament)
+        {
+            return await DbContext.Phases
+                .Where(phase => phase.TournamentId == tournament.Id)
+                .ToListAsync();
+        }
+        
+        public async Task<List<Match>> GetMatchesByPhaseAsync(Phase phase)
+        {
+            return await DbContext.Matches
+                .Where(match => match.PhaseId == phase.Id)
+                .ToListAsync();
+        }
+
         public List<Phase> GetPhasesByTournament(Tournament tournament)
         {
-            using (var db = new KandandaDbContext())
-            {
-                return db.Phases
-                    .Where(phase => phase.TournamentId == tournament.Id)
-                    .ToList();
-            }
+            return GetPhasesByTournamentAsync(tournament).Result;
         }
-        
+
         public List<Match> GetMatchesByPhase(Phase phase)
         {
-            using (var db = new KandandaDbContext())
-            {
-                return db.Matches
-                    .Where(match => match.PhaseId == phase.Id)
-                    .ToList();
-            }
+            return GetMatchesByPhaseAsync(phase).Result;
         }
-        
+
         public Phase GeneratePhase(Tournament tournament, int groupSize)
         {
             var participants = GetParticipantsByTournament(tournament);
@@ -44,14 +57,18 @@ namespace Kandanda.BusinessLayer.ServiceImplementations
             var groupPhaseGenerator = new GroupPhaseGenerator(participants, groupSize);
             var matches = groupPhaseGenerator.GenerateMatches();
 
-            var matchService = new MatchService();
-
+            var matchService = new MatchService(DbContext);
+            var phase = _phaseService.CreateEmpty();
+            
             foreach (var match in matches)
             {
+                match.PhaseId = phase.Id;
                 matchService.SaveMatch(match);
             }
 
-            var phase = new Phase();
+            phase.TournamentId = tournament.Id;
+            
+            _phaseService.Update(phase);
 
             return phase;
         }
@@ -79,18 +96,25 @@ namespace Kandanda.BusinessLayer.ServiceImplementations
                             entry.TournamentId == tournament.Id
                     select entry).FirstOrDefault();
 
-                db.TournamentParticipants.Remove(tournamentParticipant);
-                db.SaveChanges();
+                if (tournamentParticipant != null)
+                {
+                    db.TournamentParticipants.Remove(tournamentParticipant);
+                    db.SaveChanges();
+                }
             });
         }
         
+        public async Task<List<Participant>> GetParticipantsByTournamentAsync(Tournament tournament)
+        {
+            return await (from entry in DbContext.TournamentParticipants
+                join participant in DbContext.Participants
+                on entry.ParticipantId equals participant.Id
+                select participant).ToListAsync();
+        }
+
         public List<Participant> GetParticipantsByTournament(Tournament tournament)
         {
-            return ExecuteDatabaseFunc(db => 
-                (from entry in db.TournamentParticipants
-                join participant in db.Participants
-                on entry.ParticipantId equals participant.Id
-                select participant).ToList());
+            return GetParticipantsByTournamentAsync(tournament).Result;
         }
         
         public void DeleteTournament(Tournament tournament)
@@ -101,6 +125,11 @@ namespace Kandanda.BusinessLayer.ServiceImplementations
         public List<Tournament> GetAllTournaments()
         {
             return GetAll<Tournament>();
+        }
+
+        public void Update(Tournament tournament)
+        {
+            Update<Tournament>(tournament);
         }
     }
 }
